// syncro/apps/web/src/lib/api/generated/model/plantRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Syncro API
 * Syncro backend REST API.
 * OpenAPI spec version: v1
 */

export interface PlantRequest {
  /**
     * @minLength 0
     * @maxLength 64
     * @pattern [A-Za-z0-9][A-Za-z0-9_-]{0,63}
     */
  code: string;
  /**
     * @minLength 0
     * @maxLength 255
     */
  name: string;
}